Order of actions for labeled control should be focus then click
Order of actions for labeled control should be focus then click

https://bugs.webkit.org/show_bug.cgi?id=119372

Merge - https://chromium.googlesource.com/chromium/blink/+/237028d603d5fe587ec282f6d2cf3b5a31c01902

Order of actions for labeled control should be focus then click.

When we click on a checkbox or a radio button, event order is:
 - mousedown
 - focus, triggered by mousedown
 - mouseup
 - change, triggered by the following click
 - click

That is to say, we focus on it first and then click it. Click on an
associated label should have same behavior.

It is align with behavior of Blink / Chromium and Gecko (Firefox).

* Source/WebCore/html/HTMLLabelElement.cpp:
(HTMLLabelElement::defaultEventHandler) - Move dispatchSimulatedClick after "updateLayourIgnorePendingStyleSheet".
* LayoutTests/fast/forms/label/label-event-order.html: Added new test
* LayoutTests/fast/forms/label/label-event-order-expectation.txt: Added test expectations