REGRESSION (253685@main): [ Monterey wk2 ] fast/images/text-recognition/mac/cursor-types-for-recognized-text.html is a consistent failure https://bugs.webkit.org/show_bug.cgi?id=244272 <rdar://99051044>
Reviewed by Wenson Hsieh.

Test was failing because (253685@main) sets cursor:pointer for <a> with links,
since this behavior can't be obtained from cursor:auto according to specification.

Our live-text feature creates a overlay (a <div>) on the shadow DOM tree that depends on
cursor:auto behavior to render a IBeam cursor while hovering the text on top
of the image on the overlay. In this test, we had a <img> inside a <a>. After (253685@main)
the div overlay was inheriting cursor:pointer from the <a> and therefore, we would not
render a IBeam cursor for it.

Solution: We are now styling the div#image-overlay in imageOverlay.css with cursor:auto.
In that way, user can still style the wrapper image, and we preserve cursor:auto by default
for div#image-overlay.

* LayoutTests/fast/images/text-recognition/mac/cursor-types-for-recognized-text-expected.txt:
* LayoutTests/fast/images/text-recognition/mac/cursor-types-for-recognized-text.html:
testImage now tests the cursor hovering on the image area out of the text area.
testOverlayDivFromImage tests the cursor hovering the overlay div on the text area.

We have now 6 tests:
1. cursor on the image inside a link, out of text area, should be "Hand".
2. cursor on the image, with custom cursor, out of text area, should custom cursor.
3. cursor on the image, with user-select-none, inside a link, should be "Hand".
4. cursor on text area of overlay from image [1] should be "IBeam" (hinting selection).
5. cursor on text area of overlay from image [2] should be "IBeam" (hinting selection).
6. cursor on text area of overlay from image [3] should be Default (in our code called Pointer),
since [3] has user-select-none.

* Source/WebCore/html/shadow/imageOverlay.css:
(div#image-overlay):
Adding cursor:auto to div#image-overlay so we can hint text selection with IBeam cursor.

Canonical link: https://commits.webkit.org/253823@main