Disable mini mode scavenger for apps https://bugs.webkit.org/show_bug.cgi?id=244355 <rdar://98989977>
Reviewed by Yusuke Suzuki and Saam Barati.

We accidentally enabled the mini mode scavenger for Safari recently due to an entitlement
change. This is contributing to a regression in the power benchmarks, since the regular scavenger
has a period of 100 ms while the mini mode scavenger has a period of 5 ms. We should restore the old
behavior to mitigate this regression.

In talking to Geoff, it seems like he never intended for mini mode to be deployed so widely; it was
supposed to be an opt-in feature for bursty daemons that used JSC. As the mini mode appears to be
especially expensive for large apps that presumably have a lot of threads that might need to be
suspended, this patch disables the mini mode scavenger not only for Safari but for other apps as
well.

* Source/bmalloc/bmalloc/ProcessCheck.h:
(bmalloc::shouldAllowMiniMode):
* Source/bmalloc/bmalloc/ProcessCheck.mm:
(bmalloc::shouldAllowMiniMode):
* Source/bmalloc/bmalloc/bmalloc.cpp:
(bmalloc::api::enableMiniMode):

Canonical link: https://commits.webkit.org/253852@main