REGRESSION(253529@main): 2X imported/w3c/web-platform-tests/mediacapture-record/ layout tests are failing https://bugs.webkit.org/show_bug.cgi?id=244208 rdar://problem/98987676
Reviewed by Jer Noble and Eric Carlson.

When trying to convert with not enough data, the converter is blocked and does not process any future audio data.
This happened when flushing at the end and 253529@main made it happen when flushing to request data.
To workaround this, we no longer flush audio data but we make sure to let the compressor do its job.
At most 100 ms audio data might not get flushed. A follow-up should fix this.

Enable tests in mac since the underlying issue was solved a long time ago.

* LayoutTests/platform/mac-wk2/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:
* Source/WebCore/platform/mediarecorder/cocoa/AudioSampleBufferCompressor.mm:
(WebCore::AudioSampleBufferCompressor::flushInternal):

Canonical link: https://commits.webkit.org/253907@main