REGRESSION(253764@main): Disconnecting a subtree makes :lang pseudo class to never match https://bugs.webkit.org/show_bug.cgi?id=244484
Reviewed by Antti Koivisto.

The bug was caused by removedFromAncestor always clearing the effective lang in ElementRareData
when the element itself doesn't have a lang content attribute specified. This is wrong; We need
to keep the effective lang when an ancestor of "this" element still has a lang content attribute.

This patch also fixes a bug in insertedIntoAncestor that the code to inherit the effective lang
from a parent node was not running when the subtree is not inside a document or a shadow root.
To this end, insertedIntoAncestor has been refactored to match the structure of removedFromAncestor.

Finally, this patch also removes the code in insertedIntoAncestor which was trying to clear
the effective lang when "this" element doesn't have a lang attribute. This is also clearly wrong
as any ancestor with a valid lang attribute should continue to apply the same effective lang.

* LayoutTests/fast/css/lang-pseudo-disconnected-expected.txt: Added.
* LayoutTests/fast/css/lang-pseudo-disconnected.html: Added.
* Source/WebCore/dom/Element.cpp:
(WebCore::Element::insertedIntoAncestor):
(WebCore::Element::removedFromAncestor):

Canonical link: https://commits.webkit.org/253915@main