REGRESSION (iOS 16 Beta): The sound doesn't go by default through connected Airpods https://bugs.webkit.org/show_bug.cgi?id=241632 rdar://problem/96253856
Reviewed by Eric Carlson.

We were previously computing the default microphone based on the current route.
This is not always correlated with control center.
Instead, we can rely on MediaExperience AVSystemController pickableRoutesForCategory to identify which route to select as a default.
This ensures better synchronizatio with control center and does not change based on whether starting to capture on a non default audio route.
For instance, on iPad, if the default pickable route is AirPods, but the web page decides to select the built in microphone, we will no longer
consider the built in microphone as the default microphone. This will reduce the number of devicechange events as well.

Manually tested.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebCore/PAL/PAL.xcodeproj/project.pbxproj:
* Source/WebCore/PAL/pal/spi/ios/MediaExperienceAVSystemControllerSPI.h:
* Source/WebCore/platform/mediastream/ios/AVAudioSessionCaptureDeviceManager.mm:
(WebCore::computeDefaultMicrophoneInformation):
(WebCore::AVAudioSessionCaptureDeviceManager::retrieveAudioSessionCaptureDevices const):

Canonical link: https://commits.webkit.org/253921@main