Bypass call to String::substring() in HTMLConstructionSite::insertTextNode() in the common case https://bugs.webkit.org/show_bug.cgi?id=244536
Reviewed by Darin Adler.

In insertTextNode(), we only need to split the string when its length reaches
the limit, which is extremely rare. As a result, in 99.9% of cases, the call
to substring() in insertTextNode() ends up returning the same string.

This patch avoids the call to substring() entirely in the common case, since
it shows on Speedometer profiles under this call site. This is a ~0.66%
progression on Speedometer 2.1 on Apple Silicon.

* Source/WebCore/html/parser/HTMLConstructionSite.cpp:
(WebCore::HTMLConstructionSite::insertTextNode):

Canonical link: https://commits.webkit.org/253961@main