[macOS] Load PDF.js worker for optimal performance https://bugs.webkit.org/show_bug.cgi?id=244562
Reviewed by Alex Christensen.

The worker currently doesn't load because the origin for any "webkit-pdfjs-viewer" URL is "null", when it should be "webkit-pdfjs-viewer://pdfjs".

This is already the case in GTK/WPE, we just need to make the check unconditional from these platforms.

Loading the worker should be a performance improvement.

* Source/WebCore/page/SecurityOrigin.cpp:
(WebCore::shouldTreatAsUniqueOrigin):

Canonical link: https://commits.webkit.org/253969@main