Expose signpost feature through performance.mark https://bugs.webkit.org/show_bug.cgi?id=244793
Reviewed by NOBODY (OOPS!).

This patch adds signpost emission feature to performance.mark. Unfortunately, performance.mark does not include "begin" and "end"
information so we need a way to tell the interval through detail optional JS value.
We add "webkitSignpost" handling, and we carefuly retrieve a value without side effect. So while it does not break standards, still
we offer a way to tell how to emit signpost to WebKit.

    performance.mark("marking#1", { detail: { webkitSignpost: "begin" } });
    ...
    measured code.
    ...
    performance.mark("marking#1", { detail: { webkitSignpost: "end" } });

* Source/WebCore/page/PerformanceMark.cpp:
(WebCore::PerformanceMark::create):