[JSC][32bit] Enable wasm signaling memory https://bugs.webkit.org/show_bug.cgi?id=244843
Reviewed by NOBODY (OOPS!).

Enable wasm signaling memory on 32bit architectures. The only thing we
need to do is to limit the "fast mapped memory" size to 2GB, that way
we can fit that plus the redzone pages inside a 32bit pointer. This is
OK because our memory limit was already 2GB anyway
(MAX_ARRAY_BUFFER_SIZE).

* JSTests/wasm.yaml: unskip shared memory tests on arm.
* JSTests/wasm/stress/shared-memory-errors.js: ditto.
* JSTests/wasm/stress/shared-wasm-memory-buffer.js: ditto.
* Source/JavaScriptCore/wasm/WasmMemory.cpp: use a fast memory mapped
* size of 2GB for 32bit architectures.
* (JSC::Wasm::Memory::fastMappedBytes):
* Source/WTF/wtf/PlatformEnable.h: enable wasm memory signaling on
* 32bit.