[OpenSSL] Use AES_encrypt() directly for AES-CFB8 https://bugs.webkit.org/show_bug.cgi?id=244827
Reviewed by Don Olmstead.

EVP_aes_*_cfb8 are not present in boringssl. Implemen CFB8 encryption and
decryption with AES_encrypt().

In CFB8 mode, each byte of plain text is XORed with a stream cipher. This
stream cipher is generated by retrieving the first byte of an encrypted
block for each iteration. The input to the block encryption is initialized
with the initial vector and is updated by shifting the input and feeding back
the output cipher byte (XORed one) of the previous iteration.

This shift could be done by simply memmoving the input buffer every iteration.
However in this implementation we use a blocksize * 2 byte buffer and move
the pointer instead of memmoving every time..

* Source/WebCore/crypto/openssl/CryptoAlgorithmAES_CFBOpenSSL.cpp:
(WebCore::cfb8):
(WebCore::cryptEncrypt):
(WebCore::cryptDecrypt):
(WebCore::aesAlgorithm): Deleted.

Canonical link: https://commits.webkit.org/254206@main