Web Inspector: Inspector^2 can't be opened with "Allow Inspecting Web Inspector" setting enabled unless the DeveloperExtrasEnabled user default is also enabled https://bugs.webkit.org/show_bug.cgi?id=244860 rdar://87995470
Reviewed by Devin Rousso.

Since its introduction, the "Allow Inspecting Web Inspector" setting in Web Inspector hasn't worked unless you also set
the "DeveloperExtrasEnabled" default, even though that setting was being set by the Web Inspector frontend. This occured
because while the WebProcess was using the newly enabled value for the setting, the UIProcess has a one-way relationship
with the WebProcess for settings: it only propogates changes to the WebProcess, it does not itself get updated when a
preference changes in the WebProcess.

Given this, we need to inform the UIProcess when we change the setting so that both sides have a consistent view of this
state.

* Source/WebCore/inspector/InspectorFrontendClient.h:
* Source/WebCore/inspector/InspectorFrontendClientLocal.h:
* Source/WebCore/inspector/InspectorFrontendHost.cpp:
(WebCore::InspectorFrontendHost::setAllowsInspectingInspector):
* Source/WebKit/UIProcess/Inspector/RemoteWebInspectorUIProxy.cpp:
(WebKit::RemoteWebInspectorUIProxy::setInspectorPageDeveloperExtrasEnabled):
* Source/WebKit/UIProcess/Inspector/RemoteWebInspectorUIProxy.h:
* Source/WebKit/UIProcess/Inspector/RemoteWebInspectorUIProxy.messages.in:
* Source/WebKit/UIProcess/Inspector/WebInspectorUIProxy.cpp:
(WebKit::WebInspectorUIProxy::setInspectorPageDeveloperExtrasEnabled):
* Source/WebKit/UIProcess/Inspector/WebInspectorUIProxy.h:
* Source/WebKit/UIProcess/Inspector/WebInspectorUIProxy.messages.in:
* Source/WebKit/WebProcess/Inspector/RemoteWebInspectorUI.cpp:
(WebKit::RemoteWebInspectorUI::setInspectorPageDeveloperExtrasEnabled):
* Source/WebKit/WebProcess/Inspector/RemoteWebInspectorUI.h:
* Source/WebKit/WebProcess/Inspector/WebInspectorUI.cpp:
(WebKit::WebInspectorUI::setInspectorPageDeveloperExtrasEnabled):
* Source/WebKit/WebProcess/Inspector/WebInspectorUI.h:

Canonical link: https://commits.webkit.org/254236@main