Incorrect behavior of "cursor: auto" over links https://bugs.webkit.org/show_bug.cgi?id=173909 rdar://99495893
Reviewed by NOBODY (OOPS!).

The initial attempt to fix this bug and comply to spec was made
at https://commits.webkit.org/253685@main

After that, a:any-link no longer depends on 'cursor: auto'
for achieving "Hand" (css: pointer) cursor. This is because, according to spec,
'cursor: auto' can only result in default or text cursor. Therefore, 253685@main
sets 'cursor: pointer' in the UA stylesheet (html.css).

Because of that, a test case for cursor types for recognized text (live-text)
was broken. There, we were testing a <img> nested in a <a href="">. Since,
this <a> is now selected to have 'cursor: pointer', its nested <img>
has no longer 'cursor: auto', which is required for resulting in IBeam (text)
cursor for the overlay installed in the img (see EventHandler::selectCursor).

Therefore, we are updating the test case here. A <img> inside a <a href="...">
is by default styled with cursor: pointer, and therefore hovering over
its overlay doesn't show IBeam (text) cursor, unless user styles its cursor to auto.

* LayoutTests/fast/images/text-recognition/mac/cursor-types-for-recognized-text-expected.txt:
* LayoutTests/fast/images/text-recognition/mac/cursor-types-for-recognized-text.html:
Updating tests. Note that one of the test cases is commented out. As stated in the comment,
this would fail due to https://bugs.webkit.org/show_bug.cgi?id=244944 (radar: 99495849).
In short, EventHandler::selectCursor will select an IBeam cursor although it shouldn't.
This is because Node::canStartSelection() is not false for user-select: none.
* LayoutTests/platform/mac-wk2/TestExpectations:
Reseting TextExpectations for this test, before set as failure.
* Source/WebCore/html/shadow/imageOverlay.css:
(div#image-overlay):
We should not style cursor for the live-text overlay that is installed
on top of images with text. Doing that would disable user to control which
cursor is displayed for the text portion of the image, since the overlay
is in the shadow tree, hosted by the image element. I.e.: user's cursor style
for the image would not be propagated to the overlay.